api/requirements.txt
flask
requests
waitress
pillow